Move database, configuration to internal directory as it seems more appropriate structure
